+++ packages/leancode_contracts/lib/src/types/date_only.dart
@@ -1,17 +1,16 @@
 /// {@template DateOnly}
-/// Represents a single day
+/// Represents a single day.
 /// {@endtemplate}
 class DateOnly implements Comparable<DateOnly> {
-  /// Constructs a [DateOnly] from [DateTime] and asserts that it represents a valid time.
-  DateOnly(int year, int month, int day)
-      : this.fromDateTime(DateTime(year, month, day));
+  /// Constructs a [DateOnly] from `year`, `month`, and `day`.
+  /// Additionally asserts that it represents a valid date.
+  const DateOnly(this.year, this.month, this.day) : assert(year >= 0);
 
-  /// Constructs a [DateOnly] from [DateTime]
+  /// Constructs a [DateOnly] from [DateTime].
   DateOnly.fromDateTime(DateTime source)
-      : source = DateTime(source.year, source.month, source.day),
-        assert(source.year >= 0);
+      : this(source.year, source.month, source.day);
 
-  /// Deserializes a string into [DateOnly]
+  /// Deserializes a string into [DateOnly].
   factory DateOnly.fromJson(String json) {
     final parts = json.split('-');
     return DateOnly(
@@ -21,19 +20,22 @@ class DateOnly implements Comparable<DateOnly> {
     );
   }
 
-  /// [DateTime] representation of [DateOnly]
-  final DateTime source;
+  /// Year of the date.
+  final int year;
 
-  /// Year of the date
-  int get year => source.year;
+  /// Month of the year (1-indexed).
+  final int month;
 
-  /// Month of the year (1-indexed)
-  int get month => source.month;
+  /// Day of the month (1-indexed).
+  final int day;
 
-  /// Day of the month (1-indexed)
-  int get day => source.day;
+  /// Constructs a local [DateTime] representation of [DateOnly].
+  DateTime toDateTimeLocal() => DateTime(year, month, day);
 
-  /// Serializes [DateOnly] to a string.
+  /// Constructs a UTC [DateTime] representation of [DateOnly].
+  DateTime toDateTimeUtc() => DateTime.utc(year, month, day);
+
+  /// Serializes [DateOnly] into a string.
   String toJson() {
     return '${year.toString().padLeft(4, '0')}'
         '-${month.toString().padLeft(2, '0')}'
@@ -46,14 +48,28 @@ class DateOnly implements Comparable<DateOnly> {
   @override
   bool operator ==(Object other) {
     return identical(this, other) ||
-        other is DateOnly && source == other.source;
+        other is DateOnly &&
+            year == other.year &&
+            month == other.month &&
+            day == other.day;
   }
 
   @override
-  int get hashCode => source.hashCode;
+  int get hashCode => Object.hash(year, month, day);
 
   @override
   int compareTo(DateOnly other) {
-    return source.compareTo(other.source);
+    final yearCompare = year.compareTo(other.year);
+    if (yearCompare != 0) {
+      return yearCompare;
+    }
+
+    final monthCompare = month.compareTo(other.month);
+    if (monthCompare != 0) {
+      return monthCompare;
+    }
+
+    final dayCompare = day.compareTo(other.day);
+    return dayCompare;
   }
 }