+++ packages/leancode_contracts/test/types/date_only_test.dart
@@ -8,11 +8,11 @@ void main() {
   group('DateOnly', () {
     test('accepts correct dates', () {
       expect(
-        DateOnly(1999, 1, 29),
+        const DateOnly(1999, 1, 29),
         isA<DateOnly>(),
       );
-      expect(DateOnly(0, 1, 1), isA<DateOnly>());
-      expect(DateOnly(23030, 12, 31), isA<DateOnly>());
+      expect(const DateOnly(0, 1, 1), isA<DateOnly>());
+      expect(const DateOnly(23030, 12, 31), isA<DateOnly>());
     });
 
     test('drops time from DateTime', () {
@@ -36,8 +36,8 @@ void main() {
     });
 
     test('date part getters', () {
-      final one = DateOnly(1, 1, 1);
-      final complex = DateOnly(1999, 1, 29);
+      const one = DateOnly(1, 1, 1);
+      const complex = DateOnly(1999, 1, 29);
 
       expect(one.year, 1);
       expect(one.month, 1);
@@ -48,53 +48,69 @@ void main() {
       expect(complex.day, 29);
     });
 
+    test('toDateTimeLocal', () {
+      const one = DateOnly(1, 1, 1);
+      const complex = DateOnly(1999, 1, 29);
+
+      expect(one.toDateTimeLocal(), DateTime(1));
+      expect(complex.toDateTimeLocal(), DateTime(1999, 1, 29));
+    });
+
+    test('toDateTimeUtc', () {
+      const one = DateOnly(1, 1, 1);
+      const complex = DateOnly(1999, 1, 29);
+
+      expect(one.toDateTimeUtc(), DateTime.utc(1));
+      expect(complex.toDateTimeUtc(), DateTime.utc(1999, 1, 29));
+    });
+
     test('serialization', () {
-      expect(DateOnly(1, 1, 1).toJson(), '0001-01-01');
-      expect(DateOnly(1999, 1, 29).toJson(), '1999-01-29');
+      expect(const DateOnly(1, 1, 1).toJson(), '0001-01-01');
+      expect(const DateOnly(1999, 1, 29).toJson(), '1999-01-29');
     });
 
     test('deserialization', () {
-      expect(DateOnly.fromJson('0001-01-01'), DateOnly(1, 1, 1));
+      expect(DateOnly.fromJson('0001-01-01'), const DateOnly(1, 1, 1));
       expect(
         DateOnly.fromJson('1999-01-29'),
-        DateOnly(1999, 1, 29),
+        const DateOnly(1999, 1, 29),
       );
     });
 
     group('compareTo', () {
       test('greater than', () {
         expect(
-          DateOnly(2000, 1, 24).compareTo(DateOnly(2000, 1, 23)),
+          const DateOnly(2000, 1, 24).compareTo(const DateOnly(2000, 1, 23)),
           isPositive,
         );
         expect(
-          DateOnly(2000, 1, 24).compareTo(DateOnly(1999, 1, 24)),
+          const DateOnly(2000, 1, 24).compareTo(const DateOnly(1999, 1, 24)),
           isPositive,
         );
         expect(
-          DateOnly(2000, 2, 24).compareTo(DateOnly(2000, 1, 24)),
+          const DateOnly(2000, 2, 24).compareTo(const DateOnly(2000, 1, 24)),
           isPositive,
         );
       });
 
       test('less than', () {
         expect(
-          DateOnly(2000, 1, 23).compareTo(DateOnly(2000, 1, 24)),
+          const DateOnly(2000, 1, 23).compareTo(const DateOnly(2000, 1, 24)),
           isNegative,
         );
         expect(
-          DateOnly(1999, 1, 24).compareTo(DateOnly(2000, 1, 24)),
+          const DateOnly(1999, 1, 24).compareTo(const DateOnly(2000, 1, 24)),
           isNegative,
         );
         expect(
-          DateOnly(2000, 1, 24).compareTo(DateOnly(2000, 2, 24)),
+          const DateOnly(2000, 1, 24).compareTo(const DateOnly(2000, 2, 24)),
           isNegative,
         );
       });
 
       test('equal', () {
         expect(
-          DateOnly(2000, 1, 23).compareTo(DateOnly(2000, 1, 23)),
+          const DateOnly(2000, 1, 23).compareTo(const DateOnly(2000, 1, 23)),
           isZero,
         );
       });